src/main/java/yaremax/com/pb_task_24_04/service/validator/EntityValidator.java
package yaremax.com.pb_task_24_04.service.validator;

import jakarta.validation.ConstraintViolation;
import jakarta.validation.Validation;
import jakarta.validation.Validator;
import jakarta.validation.ValidatorFactory;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;

import java.util.Set;

@Component
@RequiredArgsConstructor
public class EntityValidator<T> {
    private final ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
    private final Validator validator = factory.getValidator();

    public boolean isValid(T t) {
        Set<ConstraintViolation<T>> violations = validator.validate(t);
        return violations.isEmpty();
    }
}
